Movies routes separated to api and templates routes
